+++ app/build.gradle
@@ -43,8 +43,6 @@ dependencies {
     implementation 'com.google.firebase:firebase-database:19.6.0'
     implementation 'com.google.firebase:firebase-auth:20.0.2'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'com.google.firebase:firebase-storage:20.0.0'
-    implementation 'com.squareup.picasso:picasso:2.5.2'
     //implementation 'com.google.firebase:firebase-admin:8.1.0'
 
 
@@ -53,5 +51,4 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     implementation 'com.google.firebase:firebase-bom:29.0.2'
 //    implementation 'com.google.firebase:firebase-analytics'
-
 }
