--- app/src/main/java/com/dvora/finalproject/FirebaseManager.java
@@ -1,12 +1,0 @@
-package com.dvora.finalproject;
-
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-
-public class FirebaseManager {
-    public static DatabaseReference root= FirebaseDatabase.getInstance().getReference();
-    public static FirebaseUser currentUser = FirebaseAuth.getInstance().getCurrentUser();
-
-}
