--- app/src/main/java/com/dvora/finalproject/ICallbackAdapter.java
@@ -1,7 +1,0 @@
-package com.dvora.finalproject;
-
-import com.dvora.finalproject.entities.Recipe;
-
-public interface ICallbackAdapter {
-    void onClickItem(Recipe recipe);
-}
