+++ app/src/main/java/com/dvora/finalproject/MainActivity.java
@@ -1,0 +1,43 @@
+package com.dvora.finalproject;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.Button;
+
+import com.google.firebase.auth.FirebaseAuth;
+
+public class MainActivity extends AppCompatActivity {
+    //private FirebaseAuth mAuth;
+    Button mLogoutBtn;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        //mAuth = FirebaseAuth.getInstance();
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+//        mLogoutBtn.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View view) {
+//                startActivity(new Intent(getApplicationContext(),Login.class));
+//            logout(view);
+//            }
+//        });
+        mLogoutBtn=findViewById(R.id.logoutBtn);
+        mLogoutBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                FirebaseAuth.getInstance().signOut();
+                startActivity(new Intent(getApplicationContext(),Login.class));
+                finish();
+            }
+        });
+    }
+
+//    public void logout(View view){
+//
+//    }
+}
