+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.Login">
+    tools:context=".Login">
 
     <EditText
         android:id="@+id/password"
@@ -51,8 +51,8 @@
 
     <Button
         android:id="@+id/loginBtn"
-        android:layout_width="198dp"
-        android:layout_height="44dp"
+        android:layout_width="214dp"
+        android:layout_height="54dp"
         android:layout_marginTop="76dp"
         android:text="Login"
         app:backgroundTint="#00BCD4"
