--- app/src/main/res/layout/fragment_add_recipe.xml
@@ -1,113 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#6F009688"
-    android:baselineAligned="false"
-    android:orientation="vertical"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingTop="35dp"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Add Recipe"
-        android:textSize="20dp"
-        android:textStyle="bold" />
-
-    <ScrollView
-        android:id="@+id/mainScroll"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/white">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <EditText
-                android:id="@+id/name"
-                android:layout_width="250dp"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="11dp"
-                android:hint="Name" />
-
-            <EditText
-                android:id="@+id/category"
-                android:layout_width="250dp"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="11dp"
-                android:hint="Category" />
-
-            <EditText
-                android:id="@+id/time"
-                android:layout_width="250dp"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="11dp"
-                android:hint="Time" />
-
-            <RelativeLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-                <Button
-                    android:id="@+id/add"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_alignParentRight="true"
-                    android:text="Add"
-                    app:backgroundTint="#009688" />
-
-
-                <AutoCompleteTextView
-                    android:id="@+id/textin"
-                    android:layout_width="244dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginRight="17dp"
-                    android:layout_toLeftOf="@id/quantity"
-                    android:hint="Ingredient" />
-
-                <EditText
-                    android:id="@+id/quantity"
-                    android:layout_width="50dp"
-                    android:layout_height="match_parent"
-                    android:layout_marginRight="11dp"
-                    android:layout_toLeftOf="@id/add"
-                    android:hint="gr" />
-            </RelativeLayout>
-            <!--            <ScrollView-->
-            <!--                android:layout_width="match_parent"-->
-            <!--                android:layout_height="wrap_content"-->
-            <!--                android:id="@+id/scrollView">-->
-            <LinearLayout
-                android:id="@+id/container"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical"></LinearLayout>
-
-            <EditText
-                android:id="@+id/prep"
-                android:layout_width="250dp"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="11dp"
-                android:hint="preparation method" />
-
-            <Button
-                android:id="@+id/saveBtn"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="SAVE"
-                app:backgroundTint="#009688"></Button>
-            <!--            </ScrollView>-->
-        </LinearLayout>
-
-    </ScrollView>
-
-
-</LinearLayout>
