--- app/src/main/res/layout/fragment_list_recipes.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ListRecipesFragment">
-
-    <!-- TODO: Update blank fragment layout -->
-
-    <SearchView
-        android:id="@+id/search_bar"
-        android:layout_width="wrap_content"
-        android:layout_height="37dp"
-        android:layout_below="@+id/btn_add"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="4dp"
-        android:layout_marginTop="2dp"
-        android:layout_marginEnd="39dp"
-        app:iconifiedByDefault="false"
-        app:queryHint="Search Name">
-
-    </SearchView>
-
-    <ListView
-        android:id="@+id/mainlistfragment_listv"
-        android:layout_width="match_parent"
-        android:layout_height="652dp"
-        android:layout_marginTop="80dp">
-
-    </ListView>
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="129dp"
-        android:text="My Recipes"
-        android:textAlignment="center"
-        android:textSize="30dp"
-        android:textStyle="bold"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageButton
-        android:id="@+id/btn_add"
-        android:layout_width="64dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="-3dp"
-        android:layout_marginTop="11dp"
-        android:layout_marginEnd="64dp"
-        android:layout_toStartOf="@+id/textView2"
-        android:background="#fff"
-        android:src="@android:drawable/ic_input_add" />
-
-</RelativeLayout>
