--- app/src/main/res/layout/row.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <Button
-        android:id="@+id/remove"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentRight="true"
-        android:text="Remove"
-        app:backgroundTint="#DAD9D9" />
-    <TextView
-        android:id="@+id/textout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:layout_toLeftOf="@id/textout2"/>
-    <TextView
-        android:id="@+id/textout2"
-        android:layout_width="60dp"
-        android:layout_height="wrap_content"
-        android:layout_toLeftOf="@id/remove"/>
-</RelativeLayout>
