--- app/src/main/res/menu/bottom_menu.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:title="Profile"
-        android:id="@+id/profileFragment"
-        android:icon="@drawable/profile" />
-    <item
-        android:id="@+id/listRecipesFragment2"
-        android:title="CookBook"
-        android:icon="@drawable/ic_baseline_menu_book"/>
-    <item
-        android:title="Inventory"
-        android:id="@+id/inventoryFragment"
-        android:icon="@drawable/ic_baseline_local_grocery_store"/>
-</menu>
