--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@id/listRecipesFragment2">
-
-    <fragment
-        android:id="@+id/profileFragment"
-        android:name="com.dvora.finalproject.fragments.ProfileFragment"
-        android:label="Profile"
-        tools:layout="@layout/fragment_profile" />
-    <fragment
-        android:id="@+id/listRecipesFragment2"
-        android:name="com.dvora.finalproject.ListRecipesFragment"
-        android:label="Recipes List"
-        tools:layout="@layout/fragment_list_recipes" />
-    <fragment
-        android:id="@+id/inventoryFragment"
-        android:name="com.dvora.finalproject.fragments.InventoryFragment"
-        android:label="Inventory"
-        tools:layout="@layout/fragment_inventory" />
-</navigation>
