--- app/src/main/res/values/dimen.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <dimen name="activity_vertical_margin">0dp</dimen>
-    <dimen name="activity_horizontal_margin">0dp</dimen>
-</resources>
